Analyzer/requirements.txt
paho-mqtt~=2.0.0
redis~=5.1.0b4
numpy~=2.0.0rc1